+++ src/make-detail-template.js
@@ -1,0 +1,21 @@
+export default function makeDetailTemplate(film) {
+    const dom = `            
+    <dl>
+        <dt>Title:</dt>
+        <dd>${film.title}</dd>
+        <dt>Summary:</dt>
+        <dd>${film.description}</dd>
+        <dt>Director:</dt>
+        <dd>${film.director}</dd>
+        <dt>Producer:</dt>
+        <dd>${film.producer}</dd>
+        <dt>Release Date:</dt>
+        <dd>${film.release_date}</dd>
+        <dt>RT Score:</dt>
+        <dd>${film.rt_score}</dd>
+    </dl>`;
+
+    const template = document.createElement('template');
+    template.innerHTML = dom;
+    return template.content;
+}
